src/Mask_RCNN/requirements.txt
numpy
scipy
Pillow
cython
matplotlib
scikit-image
tensorflow>=1.3.0
keras>=2.0.8
opencv-python
h5py
imgaug
IPython[all]

Package              Version
-------------------- ---------------
Cython               0.29.28
imgaug               0.4.0
matplotlib           3.3.4
numpy                1.15.3
opencv-python        3.4.5.20
Pillow               8.4.0
pip                  21.3.1
ptyprocess           0.7.0
pycocotools          2.0.4
scikit-image         0.16.2
scipy                1.5.4
tensorflow-gpu       1.13.1        2.1.0